Add files via upload
Write a program that takes a number from the user and check - if the number is greater than 30, print ""I am creating list""
if the number is less than 30 but greater than 20, print ""just joking"". if the number is less than 20 but greater than 7, print
""hello world"". If the number is less than 7 print ""arghhh, this is annoying"".