Add files via upload
Write a program that takes 3 numbers from the user. If the first number is greater than the 3rd number, print the addition
of the 3 numbers. If the 1st number is equal to the 2nd number, make 1st number equal to 3rd number and print the
message ""Kya struggle hai!"". Print the message ""question dhyan se padho""