Add files via upload
Write a program that takes a number from the user. If the number is smaller than 43, tell the user to enter a number again. Keep doing it until he enters a number greater than or equal to 43.
After getting the appropriate number, find the total number of factors of the input number except itself and print if it is an even number or odd number.